Added Ventoy as tool, as well as both big and small icons.
ShowTools ventoy,…

When installing Ventoy onto a disk with no ESP, the install script
puts the loaders into EFI/BOOT, which is not very nice.
This commit scans for /EFI/ventoy and /EFI/BOOT.
This may lead to a wrong detection but shikata nai…

Note: this was made before handle ventoy was a thing and may be
redundant. It puts ventoy in the second row with the other tools.