Avoid most screen clearing to prevent flickering.
Refind takes the safety precaution of clearing and repainting the
framebuffer, even when it's not really needed.
Clearing the screen takes some time, especially on large screens,
and this causes flickers. Painting a whole screen white, then
clear to black, then repaint to white again is not only very
disturbing, it's almost painful for the eyes.

This commit attempts to remove as many "useless" repaints as
possible to achieve a flicker-free experience from power on to
the OS. But I fear the definition of "useless" depends a lot
on the user's machine and configuration.

Note: this mod needs thorough testings. It works like a charm on
Linux with a fullscreen banner. It may not work well on some
configurations. For example, when not using a fullscreen background
image, it may be important to clear the framebuffer to remove
the OEM vendor logo. It may not work at all on Macs (I cannot test)