Only one of each tool.
The current code doesn't stop when a tool is found and adds
as many instances as possible, creating duplicates.
This can balloon very quickly when using (feat) Ventoy as a tool
(it has multiple valid loaders), or when having multiple distros
providing memtest at various places.

Note: there may be cases in the wild where having all tools is
desiderable, like having 2 different versions of the same tool,
or having 2 different tools having the same name.
Maybe this should be a config option?